CR-15: added mechanism for overriding feature flag default value for feature flag component
